move SNP and SEV-specific constructs into respective submodules
