+++ src/launch/linux/snp.rs
@@ -1,131 +1,22 @@
 // SPDX-License-Identifier: Apache-2.0
 
-//! Types for interacting with the KVM-SEV guest management API.
+//! Types for interacting with the KVM SEV-SNP guest management API.
 
-use crate::certs::sev;
-use crate::launch::*;
-use crate::launch::{Header, Measurement, Policy, Session};
+use crate::launch::snp::*;
 
 use std::marker::PhantomData;
-use std::mem::{size_of_val, MaybeUninit};
-
-/// Initialize the SEV platform context.
-#[repr(C)]
-pub struct Init;
-
-#[repr(transparent)]
-pub struct Handle(u32);
-
-impl From<LaunchStart<'_>> for Handle {
-    fn from(ls: LaunchStart) -> Self {
-        ls.handle
-    }
-}
-
-/// Initiate SEV launch flow.
-#[repr(C)]
-pub struct LaunchStart<'a> {
-    handle: Handle,
-    policy: Policy,
-    dh_addr: u64,
-    dh_len: u32,
-    session_addr: u64,
-    session_len: u32,
-    _phantom: PhantomData<&'a ()>,
-}
-
-impl<'a> LaunchStart<'a> {
-    pub fn new(policy: &'a Policy, dh: &'a sev::Certificate, session: &'a Session) -> Self {
-        Self {
-            handle: Handle(0), /* platform will generate one for us */
-            policy: *policy,
-            dh_addr: dh as *const _ as _,
-            dh_len: size_of_val(dh) as _,
-            session_addr: session as *const _ as _,
-            session_len: size_of_val(session) as _,
-            _phantom: PhantomData,
-        }
-    }
-}
-
-/// Encrypt guest data with its VEK.
-#[repr(C)]
-pub struct LaunchUpdateData<'a> {
-    addr: u64,
-    len: u32,
-    _phantom: PhantomData<&'a ()>,
-}
-
-impl<'a> LaunchUpdateData<'a> {
-    pub fn new(data: &'a [u8]) -> Self {
-        Self {
-            addr: data.as_ptr() as _,
-            len: data.len() as _,
-            _phantom: PhantomData,
-        }
-    }
-}
-
-/// Inject a secret into the guest.
-#[repr(C)]
-pub struct LaunchSecret<'a> {
-    hdr_addr: u64,
-    hdr_len: u32,
-    guest_addr: u64,
-    guest_len: u32,
-    trans_addr: u64,
-    trans_len: u32,
-    _phantom: PhantomData<&'a ()>,
-}
-
-impl<'a> LaunchSecret<'a> {
-    pub fn new(header: &'a Header, guest: usize, trans: &'a [u8]) -> Self {
-        Self {
-            hdr_addr: header as *const _ as _,
-            hdr_len: size_of_val(header) as _,
-            guest_addr: guest as _,
-            guest_len: trans.len() as _,
-            trans_addr: trans.as_ptr() as _,
-            trans_len: trans.len() as _,
-            _phantom: PhantomData,
-        }
-    }
-}
-
-/// Get the guest's measurement.
-#[repr(C)]
-pub struct LaunchMeasure<'a> {
-    addr: u64,
-    len: u32,
-    _phantom: PhantomData<&'a Measurement>,
-}
-
-impl<'a> LaunchMeasure<'a> {
-    pub fn new(measurement: &'a mut MaybeUninit<Measurement>) -> Self {
-        Self {
-            addr: measurement.as_mut_ptr() as _,
-            len: size_of_val(measurement) as _,
-            _phantom: PhantomData,
-        }
-    }
-}
-
-/// Complete the SEV launch flow and transition guest into
-/// ready state.
-#[repr(C)]
-pub struct LaunchFinish;
 
 /// Initialize the SEV-SNP platform in KVM.
 #[derive(Default)]
 #[repr(C, packed)]
-pub struct SnpInit {
+pub struct Init {
     /// Reserved space, must be always set to 0 when issuing the ioctl.
     flags: u64,
 }
 
 /// Initialize the flow to launch a guest.
 #[repr(C)]
-pub struct SnpLaunchStart<'a> {
+pub struct LaunchStart<'a> {
     /// Guest policy. See Table 7 of the AMD SEV-SNP Firmware
     /// specification for a description of the guest policy structure.
     policy: u64,
@@ -149,8 +40,8 @@ pub struct SnpLaunchStart<'a> {
     _phantom: PhantomData<&'a [u8]>,
 }
 
-impl<'a> SnpLaunchStart<'a> {
-    pub fn new(start: &'a SnpStart) -> Self {
+impl<'a> LaunchStart<'a> {
+    pub fn new(start: &'a Start) -> Self {
         let uaddr = if let Some(addr) = start.ma_uaddr {
             addr.as_ptr() as u64
         } else {
@@ -171,7 +62,7 @@ impl<'a> SnpLaunchStart<'a> {
 
 /// Insert pages into the guest physical address space.
 #[repr(C)]
-pub struct SnpLaunchUpdate<'a> {
+pub struct LaunchUpdate<'a> {
     /// guest start frame number.
     start_gfn: u64,
 
@@ -201,8 +92,8 @@ pub struct SnpLaunchUpdate<'a> {
     _phantom: PhantomData<&'a ()>,
 }
 
-impl<'a, 'b> SnpLaunchUpdate<'a> {
-    pub fn new(update: &'a SnpUpdate) -> Self {
+impl<'a, 'b> LaunchUpdate<'a> {
+    pub fn new(update: &'a Update) -> Self {
         Self {
             start_gfn: update.start_gfn,
             uaddr: update.uaddr.as_ptr() as _,
@@ -221,7 +112,7 @@ pub const KVM_SEV_SNP_FINISH_DATA_SIZE: usize = 32;
 
 /// Complete the guest launch flow.
 #[repr(C)]
-pub struct SnpLaunchFinish<'a> {
+pub struct LaunchFinish<'a> {
     /// Userspace address of the ID block. Ignored if ID_BLOCK_EN is 0.
     id_block_uaddr: u64,
 
@@ -243,8 +134,8 @@ pub struct SnpLaunchFinish<'a> {
     _phantom: PhantomData<&'a [u8]>,
 }
 
-impl<'a> SnpLaunchFinish<'a> {
-    pub fn new(finish: &'a SnpFinish) -> Self {
+impl<'a> LaunchFinish<'a> {
+    pub fn new(finish: &'a Finish) -> Self {
         let id_block = if let Some(addr) = finish.id_block {
             addr.as_ptr() as u64
         } else {
