+++ src/session/mod.rs
@@ -17,28 +17,28 @@ pub struct Initialized;
 pub struct Measuring(hash::Hasher);
 
 /// Denotes an agreeable measurement with the AMD SP.
-pub struct Verified(launch::Measurement);
+pub struct Verified(launch::sev::Measurement);
 
 /// Describes a secure channel with the AMD SP.
 ///
 /// This is required for facilitating an SEV launch and attestation.
 pub struct Session<T> {
-    policy: launch::Policy,
+    policy: launch::sev::Policy,
     tek: key::Key,
     tik: key::Key,
     data: T,
 }
 
-impl launch::Policy {
+impl launch::sev::Policy {
     fn bytes(self) -> [u8; 4] {
         unsafe { std::mem::transmute(self) }
     }
 }
 
-impl std::convert::TryFrom<launch::Policy> for Session<Initialized> {
+impl std::convert::TryFrom<launch::sev::Policy> for Session<Initialized> {
     type Error = std::io::Error;
 
-    fn try_from(value: launch::Policy) -> Result<Self> {
+    fn try_from(value: launch::sev::Policy) -> Result<Self> {
         Ok(Self {
             tek: key::Key::random(16)?,
             tik: key::Key::random(16)?,
@@ -49,7 +49,7 @@ impl std::convert::TryFrom<launch::Policy> for Session<Initialized> {
 }
 
 impl Session<Initialized> {
-    fn session(&self, nonce: [u8; 16], iv: [u8; 16], z: key::Key) -> Result<launch::Session> {
+    fn session(&self, nonce: [u8; 16], iv: [u8; 16], z: key::Key) -> Result<launch::sev::Session> {
         let master = z.derive(16, &nonce, "sev-master-secret")?;
         let kek = master.derive(16, &[], "sev-kek")?;
         let kik = master.derive(16, &[], "sev-kik")?;
@@ -71,7 +71,7 @@ impl Session<Initialized> {
         let wmac = kik.mac(&wrap)?;
         let pmac = self.tik.mac(&self.policy.bytes())?;
 
-        Ok(launch::Session {
+        Ok(launch::sev::Session {
             policy_mac: pmac,
             wrap_mac: wmac,
             wrap_tk: wrap,
@@ -81,7 +81,7 @@ impl Session<Initialized> {
     }
 
     /// Produces data needed to initiate the SEV launch sequence.
-    pub fn start(&self, chain: certs::Chain) -> Result<launch::Start> {
+    pub fn start(&self, chain: certs::Chain) -> Result<launch::sev::Start> {
         use certs::*;
 
         let pdh = chain.verify()?;
@@ -93,7 +93,7 @@ impl Session<Initialized> {
         rand::rand_bytes(&mut nonce)?;
         rand::rand_bytes(&mut iv)?;
 
-        Ok(launch::Start {
+        Ok(launch::sev::Start {
             policy: self.policy,
             cert: crt,
             session: self.session(nonce, iv, z)?,
@@ -118,7 +118,7 @@ impl Session<Initialized> {
         self,
         digest: &[u8],
         build: Build,
-        msr: launch::Measurement,
+        msr: launch::sev::Measurement,
     ) -> Result<Session<Verified>> {
         let key = pkey::PKey::hmac(&*self.tik)?;
         let mut sig = sign::Signer::new(hash::MessageDigest::sha256(), &key)?;
@@ -146,7 +146,7 @@ impl Session<Initialized> {
     /// # Safety
     ///
     /// This method must only be used in tests or unattested workflows.
-    pub unsafe fn mock_verify(self, msr: launch::Measurement) -> Result<Session<Verified>> {
+    pub unsafe fn mock_verify(self, msr: launch::sev::Measurement) -> Result<Session<Verified>> {
         Ok(Session {
             policy: self.policy,
             tek: self.tek,
@@ -166,7 +166,11 @@ impl Session<Measuring> {
     }
 
     /// Verifies the session's measurement against the AMD SP's measurement.
-    pub fn verify(mut self, build: Build, msr: launch::Measurement) -> Result<Session<Verified>> {
+    pub fn verify(
+        mut self,
+        build: Build,
+        msr: launch::sev::Measurement,
+    ) -> Result<Session<Verified>> {
         let digest = self.data.0.finish()?;
         let session = Session {
             policy: self.policy,
@@ -181,7 +185,11 @@ impl Session<Measuring> {
 
 impl Session<Verified> {
     /// Creates a packet for a secret to be injected into the guest.
-    pub fn secret(&self, flags: launch::HeaderFlags, data: &[u8]) -> Result<launch::Secret> {
+    pub fn secret(
+        &self,
+        flags: launch::sev::HeaderFlags,
+        data: &[u8],
+    ) -> Result<launch::sev::Secret> {
         let mut iv = [0u8; 16];
         rand::rand_bytes(&mut iv)?;
 
@@ -201,8 +209,8 @@ impl Session<Verified> {
         let mut mac = [0u8; 32];
         sig.sign(&mut mac)?;
 
-        Ok(launch::Secret {
-            header: launch::Header { flags, iv, mac },
+        Ok(launch::sev::Secret {
+            header: launch::sev::Header { flags, iv, mac },
             ciphertext,
         })
     }
@@ -216,7 +224,7 @@ mod initialized {
     #[test]
     fn session() {
         let session = Session {
-            policy: launch::Policy::default(),
+            policy: launch::sev::Policy::default(),
             tek: key::Key::new(vec![0u8; 16]),
             tik: key::Key::new(vec![0u8; 16]),
             data: Initialized,
@@ -266,7 +274,7 @@ mod initialized {
             0x78, 0x52, 0xb8, 0x55,
         ];
 
-        let measurement = launch::Measurement {
+        let measurement = launch::sev::Measurement {
             measure: [
                 0x6f, 0xaa, 0xb2, 0xda, 0xae, 0x38, 0x9b, 0xcd, 0x34, 0x05, 0xa0, 0x5d, 0x6c, 0xaf,
                 0xe3, 0x3c, 0x04, 0x14, 0xf7, 0xbe, 0xdd, 0x0b, 0xae, 0x19, 0xba, 0x5f, 0x38, 0xb7,
@@ -278,8 +286,8 @@ mod initialized {
             ],
         };
 
-        let policy = launch::Policy {
-            flags: launch::PolicyFlags::default(),
+        let policy = launch::sev::Policy {
+            flags: launch::sev::PolicyFlags::default(),
             minfw: Default::default(),
         };
